+++ src/Application/Masa.Alert.Application.Contracts/AlarmHistories/Dtos/GetAlarmHistoryInputDto.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the Apache License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.Alert.Application.Contracts.AlarmHistorys.Dtos;
+namespace Masa.Alert.Application.Contracts.AlarmHistories.Dtos;
 
 public class GetAlarmHistoryInputDto : PaginatedOptionsDto
 {
