+++ src/Application/Masa.Alert.Application/AlarmHistories/Queries/GetAlarmHistoryListQuery.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the Apache License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.Alert.Application.AlarmHistorys.Queries;
+namespace Masa.Alert.Application.AlarmHistories.Queries;
 
 public record GetAlarmHistoryListQuery(GetAlarmHistoryInputDto Input) : Query<PaginatedListDto<AlarmHistoryDto>>
 {
