+++ src/Application/Masa.Alert.Application/_Imports.cs
@@ -33,12 +33,12 @@ global using Masa.BuildingBlocks.StackSdks.Mc;
 global using Masa.BuildingBlocks.StackSdks.Auth.Contracts.Model;
 global using Masa.BuildingBlocks.StackSdks.Mc.Enum;
 global using Masa.BuildingBlocks.StackSdks.Mc.Model;
-global using Masa.Alert.Domain.AlarmHistorys.Repositories;
-global using Masa.Alert.Application.Contracts.AlarmHistorys.Dtos;
+global using Masa.Alert.Domain.AlarmHistories.Repositories;
+global using Masa.Alert.Application.Contracts.AlarmHistories.Dtos;
 global using Masa.Alert.Application.AlarmRules.Queries;
 global using Masa.Alert.Application.Contracts.QueryContext;
-global using Masa.Alert.Domain.AlarmHistorys.Aggregates;
+global using Masa.Alert.Domain.AlarmHistories.Aggregates;
 global using Microsoft.EntityFrameworkCore.Query;
 global using Masa.Alert.Application.Contracts.QueryModels;
 global using Masa.Alert.Domain.Shared.AlarmHistory;
-global using static Masa.Alert.Application.Contracts.AlarmHistorys.Dtos.GetAlarmHistoryInputDto;
+global using static Masa.Alert.Application.Contracts.AlarmHistories.Dtos.GetAlarmHistoryInputDto;
