+++ src/Domain/Masa.Alert.Domain.Shared/AlarmHistories/AlarmHistoryStatuses.cs
