+++ src/Domain/Masa.Alert.Domain/AlarmHistories/EventHandler/TriggerAlarmEventHandler.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the Apache License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.Alert.Domain.AlarmHistorys.EventHandler;
+namespace Masa.Alert.Domain.AlarmHistories.EventHandler;
 
 public class TriggerAlarmEventHandler
 {
