+++ src/Domain/Masa.Alert.Domain/AlarmHistories/Events/RecoveryAlarmEvent.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the Apache License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.Alert.Domain.AlarmHistorys.Events;
+namespace Masa.Alert.Domain.AlarmHistories.Events;
 
 public record RecoveryAlarmEvent(Guid AlarmRuleId) : DomainEvent
 {
