+++ src/Domain/Masa.Alert.Domain/_Imports.cs
@@ -21,11 +21,11 @@ global using FluentValidation.Results;
 global using Masa.Contrib.Dispatcher.Events;
 global using Microsoft.Extensions.Logging;
 global using Masa.Alert.Domain.Shared.AlarmHistory;
-global using Masa.Alert.Domain.AlarmHistorys;
-global using Masa.Alert.Domain.AlarmHistorys.Events;
+global using Masa.Alert.Domain.AlarmHistories;
+global using Masa.Alert.Domain.AlarmHistories.Events;
 global using Masa.Alert.Domain.Shared;
-global using Masa.Alert.Domain.AlarmHistorys.Aggregates;
-global using Masa.Alert.Domain.AlarmHistorys.Repositories;
+global using Masa.Alert.Domain.AlarmHistories.Aggregates;
+global using Masa.Alert.Domain.AlarmHistories.Repositories;
 global using Masa.BuildingBlocks.ReadWriteSplitting.Cqrs.Commands;
 global using Masa.Alert.Infrastructure.Common.Utils;
 global using System.Threading.Channels;
