+++ src/Infrastructure/Masa.Alert.EntityFrameworkCore/_Imports.cs
@@ -14,7 +14,7 @@ global using System.Threading.Channels;
 global using System.Collections.Concurrent;
 global using Masa.Alert.Domain.AlarmRules.Aggregates;
 global using Masa.Alert.Domain.AlarmRules.Repositories;
-global using Masa.Alert.Domain.AlarmHistorys.Aggregates;
-global using Masa.Alert.Domain.AlarmHistorys.Repositories;
+global using Masa.Alert.Domain.AlarmHistories.Aggregates;
+global using Masa.Alert.Domain.AlarmHistories.Repositories;
 global using Masa.Alert.Application.Contracts.QueryContext;
 global using Masa.Alert.Application.Contracts.QueryModels;
