+++ src/Services/Masa.Alert.Service/_Imports.cs
@@ -77,7 +77,7 @@ global using Masa.Alert.Application.AlarmRules.Commands;
 global using Masa.Contrib.Caching.Distributed.StackExchangeRedis;
 global using Masa.Alert.Application.Contracts.QueryContext;
 global using Microsoft.Extensions.DependencyInjection;
-global using Masa.Alert.Application.AlarmHistorys.Queries;
-global using Masa.Alert.Application.Contracts.AlarmHistorys.Dtos;
+global using Masa.Alert.Application.AlarmHistories.Queries;
+global using Masa.Alert.Application.Contracts.AlarmHistories.Dtos;
 global using Masa.Alert.Domain.Shared.AlarmHistory;
-global using static Masa.Alert.Application.Contracts.AlarmHistorys.Dtos.GetAlarmHistoryInputDto;
+global using static Masa.Alert.Application.Contracts.AlarmHistories.Dtos.GetAlarmHistoryInputDto;
