+++ src/Web/Masa.Alert.Web.Admin/_Imports.cs
@@ -43,7 +43,7 @@ global using Masa.Alert.Domain.Shared.AlarmRules;
 global using static Masa.Stack.Components.JsInitVariables;
 global using Masa.Alert.Web.Admin.ViewModel.Enums;
 global using Masa.Alert.Infrastructure.Common.Utils;
-global using Masa.Alert.Application.Contracts.AlarmHistorys.Dtos;
+global using Masa.Alert.Application.Contracts.AlarmHistories.Dtos;
 global using Masa.Alert.Application.Contracts.AlarmRules.Dtos;
 global using Masa.Alert.Domain.Shared.AlarmHistory;
 global using Masa.Alert.Web.Admin.ViewModel.AlarmHistory;
@@ -69,6 +69,6 @@ global using Masa.BuildingBlocks.StackSdks.Pm.Model;
 global using Masa.BuildingBlocks.StackSdks.Tsc;
 global using Masa.Alert.Contracts.Admin.Tsc;
 global using Masa.BuildingBlocks.StackSdks.Tsc.Model;
-global using Masa.Alert.ApiGateways.Caller.Services.AlarmHistorys;
+global using Masa.Alert.ApiGateways.Caller.Services.AlarmHistories;
 global using Masa.Alert.Web.Admin.Components.Modules.Alarm.Model;
 global using Masa.BuildingBlocks.StackSdks.Mc;
