+++ src/Web/Masa.Alert.Web.Admin/_Imports.razor
@@ -32,7 +32,7 @@
 @using Masa.Alert.Web.Admin.Pages.AlarmHistory.Modules
 @using Masa.Alert.Domain.Shared.AlarmHistory
 @using static Masa.Alert.Application.Contracts.AlarmRules.Dtos.GetAlarmRuleInputDto
-@using static Masa.Alert.Application.Contracts.AlarmHistorys.Dtos.GetAlarmHistoryInputDto
+@using static Masa.Alert.Application.Contracts.AlarmHistories.Dtos.GetAlarmHistoryInputDto
 @using Masa.Alert.Web.Admin.Pages.WebHooks.Modules
 @using Masa.Alert.Contracts.Admin.Tsc
 @using Masa.Alert.Web.Admin.Components.Modules.Alarm.Model
